src/test/java/com/springframework/springrecipeapp/services/ImageServiceImplTest.java
package com.springframework.springrecipeapp.services;

import com.springframework.springrecipeapp.model.Recipe;
import com.springframework.springrecipeapp.repository.RecipeRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.ArgumentCaptor;
import org.mockito.Mock;
import org.springframework.mock.web.MockMultipartFile;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.web.multipart.MultipartFile;

import java.io.IOException;
import java.util.Optional;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

class ImageServiceImplTest {

    @Mock
    RecipeRepository repository;

    ImageService imageService;

//    MockMvc mockMvc;

    @BeforeEach
    void setUp() {
        imageService = new ImageServiceImpl(repository);
//        mockMvc = MockMvcBuilders.standaloneSetup()

    }

    @Test
    void saveImageFile() throws IOException {
        Long id = 1L;
        MultipartFile multipartFile = new MockMultipartFile(
                "imagefile", "testing.txt", "text/plain", "Spring framework".getBytes());

        Recipe recipe = new Recipe();

        recipe.setId(id);
        Optional<Recipe> recipeOptional = Optional.of(recipe);
        when(repository.findById(anyLong())).thenReturn(recipeOptional);

        ArgumentCaptor<Recipe> argumentCaptor = ArgumentCaptor.forClass(Recipe.class);

        //when
        imageService.saveImageFile(id, multipartFile);

        //then
        verify(repository).save(argumentCaptor.capture());
        var savedRecipe = argumentCaptor.getValue();
        assertEquals(multipartFile.getBytes().length, savedRecipe.getImage().length);

    }
}